Shvaan/requirements.txt
asgiref==3.5.2
backports.cached-property==1.0.2
boto3==1.24.2
botocore==1.27.33
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.1
cryptography==37.0.2
Django==4.0.5
django-ipware==4.0.2
django-otp==1.1.3
django-phonenumber-field==6.3.0
django-storages==1.12.3
environs==9.5.0
factory-boy==3.2.1
Faker==13.12.0
flake8==4.0.1
graphql-core==3.2.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
marshmallow==3.17.0
mccabe==0.6.1
model-bakery==1.7.0
packaging==21.3
phonenumbers==8.12.54
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
pymemcache==3.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
Pillow==9.3.0
requests==2.27.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
strawberry-graphql==0.114.0
typing_extensions==4.3.0
urllib3==1.26.10
django-tinymce==3.5.0
